refactor: context-managed dependencies
